Renamed `bfrt_info` to bfrt and removed the `bfrt` prefix of the sub-files
